requirements.txt
fastapi
uvicorn
matplotlib
pydantic
python-dotenv
numpy
